# Support specializing get_hash without nightly