--- a/Cargo.toml
+++ b/Cargo.toml
@@ -85,6 +85,7 @@ cfg-if = "1.0"
 portable-atomic = { version = "1.0.0", optional = true }
 getrandom = { version = "0.2.7", optional = true }
 zerocopy = { version = "0.7.31", default-features = false, features = ["simd"] }
+typeid = "1.0"
 
 [target.'cfg(not(all(target_arch = "arm", target_os = "none")))'.dependencies]
 once_cell = { version = "1.18.0", default-features = false, features = ["alloc"] }