--- a/build.rs
+++ b/build.rs
@@ -7,9 +7,6 @@ fn main() {
     if let Some(true) = version_check::is_min_version("1.71.0") {
         println!("cargo:rustc-cfg=build_hasher_hash_one");
     }
-    if let Some(true) = version_check::supports_feature("specialize") {
-        println!("cargo:rustc-cfg=feature=\"specialize\"");
-    }
     let arch = env::var("CARGO_CFG_TARGET_ARCH").expect("CARGO_CFG_TARGET_ARCH was not set");
     if arch.eq_ignore_ascii_case("x86_64")
         || arch.eq_ignore_ascii_case("aarch64")