@dtolnay 2024-10-19
I agree with https://github.com/tkaitchuck/aHash/pull/242#issuecomment-2316468766 &mdash; I was hoping to get this in an 0.8 patch.