@tkaitchuck 2025-04-24
I want to pull the generic type from the type of the map and type the builder to it rather than use that parameter. The advantage of this is we don't depend on having an implementation for the dereferenced type.