@dtolnay 2025-05-04
- Rebased over #261