@dtolnay 2025-05-06
- Rebased over https://github.com/tkaitchuck/aHash/pull/269